pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[project]
name = "cellspec"
version = "0.1.0"
description = "A python package for analyzing variant calls from high throughput single cell genome sequencing experiments"
readme = "README.md"
license = { file = "LICENSE" }
maintainers = [
  { name = "Dustin Mullaney", email = "dmulla@uw.edu" },
]
authors = [
  { name = "Dustin Mullaney" },
]
requires-python = ">=3.8"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "numpy>=1.20.0",
  "pandas>=1.3.0",
  "matplotlib>=3.4.0",
  "anndata>=0.8.0",
  "scanpy>=1.9.0",
  "scipy>=1.7.0",
  "cyvcf2>=0.30.0",
  "pyfaidx>=0.7.0",
  "tqdm>=4.62.0",
  "biopython>=1.79",
  # for debug logging (referenced from the issue template)
  "session-info2",
]
# https://docs.pypi.org/project_metadata/#project-urls
urls.Documentation = "https://cellspec.readthedocs.io/"
urls.Homepage = "https://github.com/harrispopgen/cellspec"
urls.Source = "https://github.com/harrispopgen/cellspec"

[dependency-groups]
dev = [
  "pre-commit",
  "twine>=4.0.2",
]
test = [
  "coverage>=7.10",
  "pytest",
  "pytest-cov",     # For VS Code’s coverage functionality
]
doc = [
  "ipykernel",
  "ipython",
  "myst-nb>=1.1",
  "pandas",
  "sphinx>=8.1",
  "sphinx-autodoc-typehints",
  "sphinx-book-theme>=1",
  "sphinx-copybutton",
  "sphinx-design",
  "sphinxcontrib-bibtex>=1",
  "sphinxcontrib-katex",
  "sphinxext-opengraph",
]

[tool.hatch.envs.default]
installer = "uv"
dependency-groups = [ "dev" ]

[tool.hatch.envs.docs]
dependency-groups = [ "doc" ]
scripts.build = "sphinx-build -M html docs docs/_build -W {args}"
scripts.open = "python -m webbrowser -t docs/_build/html/index.html"
scripts.clean = "git clean -fdX -- {args:docs}"

# Test the lowest and highest supported Python versions with normal deps
[[tool.hatch.envs.hatch-test.matrix]]
deps = [ "stable" ]
python = [ "3.11", "3.14" ]

# Test the newest supported Python version also with pre-release deps
[[tool.hatch.envs.hatch-test.matrix]]
deps = [ "pre" ]
python = [ "3.14" ]

[tool.hatch.envs.hatch-test]
dependency-groups = [ "dev", "test" ]

[tool.hatch.envs.hatch-test.overrides]
# If the matrix variable `deps` is set to "pre",
# set the environment variable `UV_PRERELEASE` to "allow".
matrix.deps.env-vars = [
  { key = "UV_PRERELEASE", value = "allow", if = [ "pre" ] },
]

[tool.ruff]
line-length = 120
src = [ "src" ]
extend-include = [ "*.ipynb" ]

format.docstring-code-format = true

lint.select = [
  "B",      # flake8-bugbear
  "BLE",    # flake8-blind-except
  "C4",     # flake8-comprehensions
  "D",      # pydocstyle
  "E",      # Error detected by Pycodestyle
  "F",      # Errors detected by Pyflakes
  "I",      # isort
  "RUF100", # Report unused noqa directives
  "TID",    # flake8-tidy-imports
  "UP",     # pyupgrade
  "W",      # Warning detected by Pycodestyle
]
lint.ignore = [
  "B008", # Errors from function calls in argument defaults. These are fine when the result is immutable.
  "D100", # Missing docstring in public module
  "D104", # Missing docstring in public package
  "D105", # __magic__ methods are often self-explanatory, allow missing docstrings
  "D107", # Missing docstring in __init__
  # Disable one in each pair of mutually incompatible rules
  "D203", # We don’t want a blank line before a class docstring
  "D213", # <> We want docstrings to start immediately after the opening triple quote
  "D400", # first line should end with a period [Bug: doesn’t work with single-line docstrings]
  "D401", # First line should be in imperative mood; try rephrasing
  "E501", # line too long -> we accept long comment lines; formatter gets rid of long code lines
  "E731", # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
  "E741", # allow I, O, l as variable names -> I is the identity matrix
]
lint.per-file-ignores."*/__init__.py" = [ "F401" ]
lint.per-file-ignores."docs/*" = [ "I" ]
lint.per-file-ignores."tests/*" = [ "D" ]
lint.pydocstyle.convention = "numpy"

[tool.pytest]
strict = true
testpaths = [ "tests" ]
addopts = [
  "--import-mode=importlib", # allow using test files with same name
]

[tool.coverage.run]
source = [ "cellspec" ]
patch = [ "subprocess" ]
omit = [
  "**/test_*.py",
]

[tool.cruft]
skip = [
  "tests",
  "src/**/__init__.py",
  "src/**/basic.py",
  "docs/api.md",
  "docs/changelog.md",
  "docs/references.bib",
  "docs/references.md",
  "docs/notebooks/example.ipynb",
]
